+++ cm-mlops/CHANGES.md
@@ -1,5 +1,6 @@
 * 20221010:
  * added cm run script --debug-script-tags to run cmd/bash before native script
+ * added cm run script --shell to set env and run shell after script execution
 
 * 20221007:
  * added script template (used when adding new scripts)